Fix everything.
Refactor workers to not inherit 'EventEmitter'. Add a 'time' constants
module. Refactor 'MatchHistoryWorker' so that it persist the received
matches through the recursion, this will allow us to spawn the
'MatchDetailWorkers' after we have figured out the total amount of new
matches.
Add a 'PlayerSchema' that is used by 'PlayerSummaryWorker'. Implement
'MatchDetailsWorker' module.